move all topbar to the sidebar
